fix: runtime compatibility — resample, silero cache, platform warnings
- Replace scipy.resample_poly with torchaudio.functional.resample
  (scipy hangs on some envs for long audio 48kHz→16kHz)
- silero_vad: load model from local cache first, avoids network
  timeout on air-gapped servers
- faster_whisper_asr: add macOS ARM64 deadlock warning in docstring
- vkit operators show: highlight warning blocks in yellow

+++ src/voxkitchen/operators/quality/bandwidth_estimate.py
@@ -78,8 +78,8 @@ class BandwidthEstimateOperator(Operator):
         cutoff = (torch.log(ratios + 1e-12) > 12).to(torch.int).argmax()
 
         if cutoff > 0:
-            bw_khz = round(freq[cutoff].item() * 2 / 1000, 2)
+            bw_khz = float(round(freq[cutoff].item() * 2 / 1000, 2))
         else:
-            bw_khz = round(sr / 1000, 2)
+            bw_khz = float(round(sr / 1000, 2))
 
         return cut.model_copy(update={"metrics": {**cut.metrics, "bandwidth_khz": bw_khz}})
